# src/mdm.api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "行動裝置管理系統 API — 裝置管控、資產管理、借用流程、App 派發",
        "title": "MDM Server API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/device-apps": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "查詢裝置已安裝 App",
                "parameters": [
                    {
                        "type": "string",
                        "description": "裝置 UDID",
                        "name": "device_udid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{device_apps: [...]}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
        "/api/device-apps/install": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "向 MDM 發送 InstallApplication 指令，VPP App 會先指派授權",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "安裝 App 到裝置",
                "parameters": [
                    {
                        "description": "App ID + 裝置 UDID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.swagAppActionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagCommandResp"
                        }
                    },
                    "404": {
                        "description": "App 不存在",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    },
                    "409": {
                        "description": "已安裝或超過授權數量",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
        "/api/device-apps/uninstall": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "移除裝置上的 App",
                "parameters": [
                    {
                        "description": "App ID + 裝置 UDID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.swagAppActionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagCommandResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
        "/api/device-apps/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "更新裝置上的 App",
                "parameters": [
                    {
                        "description": "App ID + 裝置 UDID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.swagAppActionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagCommandResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
...
        "/api/itunes-lookup": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "iTunes Lookup（代理 Apple API）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bundle ID",
                        "name": "bundleId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "iTunes Store ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/itunes-search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "iTunes 搜尋（代理 Apple API）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜尋關鍵字",
                        "name": "term",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "結果數量上限",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/managed-apps": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "受管 App 列表 / 新增受管 App",
                "parameters": [
                    {
                        "description": "新增 App（POST）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.swagManagedAppReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET: {apps: [...]}, POST: {id: ...}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "重複",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "受管 App 列表 / 新增受管 App",
                "parameters": [
                    {
                        "description": "新增 App（POST）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.swagManagedAppReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET: {apps: [...]}, POST: {id: ...}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "重複",
                        "schema": {
                            "$ref": "#/definitions/controller.swagError"
                        }
                    }
                }
            }
        },
        "/api/managed-apps/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "更新 / 刪除受管 App",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新欄位（PUT）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagOK"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "更新 / 刪除受管 App",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新欄位（PUT）",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagOK"
                        }
                    }
                }
            }
        },
...
        "/api/sync-device-apps": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "同步所有裝置已安裝 App 到資料庫",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.swagSyncAppsResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.swagAppActionReq": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "udid": {
                    "type": "string"
                }
            }
        },
...
        "controller.swagCommandResp": {
            "type": "object",
            "properties": {
                "command_uuid": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "raw_response": {}
            }
        },
...
        "controller.swagError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "controller.swagOK": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "controller.swagSyncAppsResp": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "synced": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}